DistrictService add test unit

+++ api/src/main/java/com/edburguer/dto/DistrictDto.java
@@ -18,5 +18,5 @@ public class DistrictDto {
     private String name;
 
     @NotNull
-    private Float deliveryPrice;
+    private Double deliveryPrice;
 }
